+++ CHANGELOG.md
@@ -41,6 +41,7 @@ Ref: https://keepachangelog.com/en/1.0.0/
 - (evm) [#3](https://github.com/dymensionxyz/ethermint/pull/3) Disabled contract creation
 - (evm) [#11](https://github.com/dymensionxyz/ethermint/pull/11) Disabled contract creation for Dymension chains only (adjust the absolute logic in [#3](https://github.com/dymensionxyz/ethermint/pull/3))
 - (evm) [#11](https://github.com/dymensionxyz/ethermint/pull/11) Virtual Frontier Contract
+- (evm) [#15](https://github.com/dymensionxyz/ethermint/pull/15) VFBC deployment no longer revert changes on error while contract deployment
 
 ### Bug Fixes
 - (rpc) [#6](https://github.com/dymensionxyz/ethermint/pull/6) Zero gas config for tracing methods